Rename sub-package to "helpers" as tools.py already exists in ansys/dpf/post/
